moved class to utils; specified type hints
